cleanup

+++ site-change-notifier.sh
@@ -1,25 +1,42 @@
 #!/bin/bash
 
+#verifies correct number of arguments
+if [ $# == 0 ]; then
+	printf 'No arguments given, 2 expected.\n' >&2;
+	printf 'Usage: arg1 = site URL, arg2 = receiver email address, arg3 = sender email address\n' >&2;
+	exit 1;
+elif [ $# -lt 3 ]; then
+	printf 'Not enough arguments given, 3 expected.\n' >&2;
+	exit 1;
+elif [ $# -gt 3 ]; then
+	printf 'Too many arguments given, 3 expected.\n' >&2;
+	exit 1;
+fi
+
 #files to compare
-F1=path/to/original
-F2=path/to/new
+F1=original
+F2=new
 
-#website for wget
-site="example.com"
+#website
+site=$1
 
-#email information
-notify1="receivers@email.address"
-from="From:senders@email.address"
+#email info
+notify1=$2
+from="From:$3"
 
-#grabbing the site
-(`wget -O "$F2" "$site"`)
-
-#comparing the files
-compare=(`diff "$F1" "$F2"`)
-
-if [ "$compare" != "" ]; then
-	(`echo 'A change has been made to '$site'' | mail -s "SITE CHANGE DETECTED" "$notify1" -a "$from"`)
-	echo "change made"
+if [ -f "$F1" ]; then
+	#grabbing the site
+	(`wget -O "$F2" "$site"`)
+	
+	#compare the files
+	compare=(`diff "$F1" "$F2"`)
+	
+	if [ "$compare" != "" ]; then
+       		(`echo 'A change has been made to '$site'' | mail -s "SITE CHANGE DETECTED" "$notify1" -a "$from"`) 
+		echo "change made" 
+	else
+       		echo "no changes made" 
+	fi
 else
-	echo "no changes made"
+	(`wget -O "$F1" "$site"`)
 fi
